requirements.txt
azure-ai-vision-imageanalysis==1.0.0
azure-cognitiveservices-vision-customvision==3.1.1
azure-common==1.1.28
azure-core==1.37.0
azure-mgmt-core==1.6.0
blinker==1.6.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.6
Flask==3.0.0
flask-cors==6.0.2
gunicorn==21.2.0
idna==3.11
isodate==0.7.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msrest==0.7.1
oauthlib==3.3.1
packaging==23.2
requests==2.32.5
requests-oauthlib==2.0.0
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.0.0
